Run R worker job

Spin up an AWS EC2 instance and run an R script or R Markdown document in that instance. It launched an EC2 instance, attaches EFS storage, then send a command to execute the R script or RMD file as an SSM command. This function will block the R session until execution is complete.

run_job(
  script,
  job_id,
  image_id,
  instance_type,
  spot,
  spot_max_price,
  efs_name,
  security_group_id,
  ssm_iam_profile,
  efs_mnt_pt,
  region,
  instance_max_hours,
  exec_timeout_hours,
  r_version,
  ssm_work_dir,
  log_file,
  rmd_opts = list(),
  status_sleep_seconds = 30
)

Arguments

script
Path to the R script or R Markdown document to execute (on EFS storage)

job_id
Job id

image_id
id of the AMI to launch

instance_type
Instance type

spot
Boolean value for whether to launch a spot instance

spot_max_price
If launching a spot instance, the maximum bid price

efs_name
EFS id

security_group_id
Security group id

ssm_iam_profile
Name of IAM profile with SSM permissions. This profile will be attached to the launched EC2 instance.

efs_mnt_pt
Mount point of EFS storage on worker instance

region
AWS EC2 region

instance_max_hours
Maximum lifetime of the instance in hours before it will automatically shut down

exec_timeout_hours
Number of hours for timeout of the SSM command for executing the script

r_version
Version of R to use

ssm_work_dir
Working directory of SSM command for executing R script/document

log_file
Path of log file to which stdout/stderr of R script will be written

rmd_opts
If running an R Markdown document, optional parameters for the output file path and document parameters

status_sleep_seconds
How many seconds to wait between checking the status of the execution of the R script